test: move starknetid tests in a separate file
